move create_default_context out of cli.rs and into its own mod
