+++ crates/nu-cli/src/commands.rs
@@ -30,6 +30,7 @@ pub(crate) mod date;
 pub(crate) mod debug;
 pub(crate) mod def;
 pub(crate) mod default;
+pub(crate) mod default_context;
 pub(crate) mod describe;
 pub(crate) mod do_;
 pub(crate) mod drop;
